+++ typeid/typeid-go/encoding_test.go
@@ -44,15 +44,3 @@ func TestJSON_Subtype(t *testing.T) {
 	err = json.Unmarshal(encoded, &wrongType)
 	assert.Error(t, err)
 }
-
-func TestScan(t *testing.T) {
-	str := "prefix_00041061050r3gg28a1c60t3gf"
-	tid := typeid.Must(typeid.FromString(str))
-
-	var scanned typeid.AnyID
-	err := scanned.Scan(str)
-	assert.NoError(t, err)
-
-	assert.Equal(t, tid, scanned)
-	assert.Equal(t, str, scanned.String())
-}
